Implement parallel formulae reading
